+++ extensions/zen-browser/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Added Open Workspace Command] - 2025-09-14
 
+- Added a command to open a specific workspace in Zen Browser.
+- Added a hook to fetch user shortcuts.
+
 ## [Fix New Tab and History Tab] - 2025-06-11
 
 - Fix new tab and history tab commands, by adding a delay to the script to ensure the URL bar is focused.